Refactor: add popupForm component and use it in VoteForm and NeedForm
