+++ src/app/components/Forms/VoteForm.tsx
@@ -8,6 +8,7 @@ import {
   FaRegQuestionCircle,
   FaExclamationCircle,
 } from 'react-icons/fa'
+import PopupForm from '../Forms/PopupForm'
 
 type VoteFormProps = {
   onSubmit: (payload: { type: VoteTypes; text: string }) => void
@@ -65,11 +66,11 @@ export default function VoteForm({
   }
 
   return (
-    <FormContainer
+    <PopupForm
+      SubmitText="Add Vote"
       onSubmit={handleSubmit}
-      color={voteTypeColors[voteType as keyof typeof voteTypeColors]}
+      onCancel={handleCancel}
     >
-      <FormTitle>Add Vote</FormTitle>
       <div>
         <h4>Consent Level</h4>
         <VoteTypePicker>
@@ -97,45 +98,9 @@ export default function VoteForm({
         }}
       />
       {submitError && <p>You have to choose a consent level before voting.</p>}
-      <ButtonContainer>
-        <Button type="button" onClick={handleCancel}>
-          Cancel
-        </Button>
-        <Button>Vote</Button>
-      </ButtonContainer>
-    </FormContainer>
+    </PopupForm>
   )
 }
-
-const FormTitle = styled.h2`
-  font-size: 2.8rem;
-  text-transform: uppercase;
-`
-
-const FormContainer = styled.form<{ color: string }>`
-  position: absolute;
-  bottom: 0;
-  right: 0;
-  left: 0;
-  display: flex;
-  flex-direction: column;
-  gap: 20px;
-  padding: 20px;
-  color: rgba(0, 0, 0, 0.4);
-  background-color: ${({ color }) => color};
-  border-top-right-radius: 30px;
-  border-top-left-radius: 30px;
-  box-shadow: rgba(0, 0, 0, 0.25) 0px 54px 55px,
-    rgba(0, 0, 0, 0.12) 0px -12px 30px, rgba(0, 0, 0, 0.12) 0px 4px 6px,
-    rgba(0, 0, 0, 0.17) 0px 12px 13px, rgba(0, 0, 0, 0.09) 0px -3px 5px;
-  transition: 0.2s;
-`
-
-const ButtonContainer = styled.div`
-  display: grid;
-  grid-template-columns: 1fr 1fr;
-  gap: 20px;
-`
 
 const VoteTypePicker = styled.fieldset`
   display: flex;
